changed files to proper packages
