+++ src/main/java/com/capgemini/molvenoresort/enums/BedType.java
@@ -1,4 +1,4 @@
-package com.capgemini.molvenoresort.room;
+package com.capgemini.molvenoresort.enums;
 
 public enum BedType {
     SINGLE, DOUBLE, CHILD, BABY
