+++ src/main/java/com/capgemini/molvenoresort/enums/RoomStatus.java
@@ -1,4 +1,4 @@
-package com.capgemini.molvenoresort.room;
+package com.capgemini.molvenoresort.enums;
 
 public enum RoomStatus {
 	AVAILABLE, IN_CLEANING, UNAVAILABLE
