+++ src/main/java/com/capgemini/molvenoresort/enums/RoomType.java
@@ -1,4 +1,4 @@
-package com.capgemini.molvenoresort.room;
+package com.capgemini.molvenoresort.enums;
 
 public enum  RoomType {
     SINGLE, DOUBLE, DOUBLEX_2, PENTHOUSE,
