+++ src/main/java/com/capgemini/molvenoresort/models/Employee.java
@@ -1,9 +1,8 @@
-package com.capgemini.molvenoresort.user;
+package com.capgemini.molvenoresort.models;
+
+import com.capgemini.molvenoresort.enums.EmployeeRole;
 
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.validation.constraints.NotNull;
 import java.util.Date;
 import java.util.Objects;
