+++ src/main/java/com/capgemini/molvenoresort/repositories/EmployeeRepository.java
@@ -1,5 +1,6 @@
-package com.capgemini.molvenoresort.user;
+package com.capgemini.molvenoresort.repositories;
 
+import com.capgemini.molvenoresort.models.Employee;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
