+++ src/main/java/com/capgemini/molvenoresort/repositories/bookingRepository.java
@@ -1,5 +1,6 @@
-package com.capgemini.molvenoresort.booking;
+package com.capgemini.molvenoresort.repositories;
 
+import com.capgemini.molvenoresort.models.Booking;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
