--- src/main/java/com/capgemini/molvenoresort/user/EmployeeController.java
@@ -1,33 +1,0 @@
-package com.capgemini.molvenoresort.user;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-@RestController
-@RequestMapping("/employees")
-public class EmployeeController {
-
-	@Autowired
-	private EmployeeRepository employeeRepository;
-
-	@GetMapping
-	public List<Employee> getEmployees() {
-		return (ArrayList)employeeRepository.findAll();
-	}
-
-	@GetMapping("/{id}")
-	public ResponseEntity<Employee> getEmployeeById(@PathVariable long id) {
-		Optional<Employee> result = employeeRepository.findById(id);
-		return result.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
-	}
-
-	@PostMapping("/add")
-	public ResponseEntity<Employee> addEmployee(@RequestBody Employee employee) {
-		return ResponseEntity.ok(employeeRepository.save(employee));
-	}
-}
