+++ src/test/java/com/capgemini/molvenoresort/room/RoomControllerTest.java
@@ -1,10 +1,12 @@
 package com.capgemini.molvenoresort.room;
 
+import com.capgemini.molvenoresort.controllers.RoomController;
+import com.capgemini.molvenoresort.enums.RoomType;
+import com.capgemini.molvenoresort.models.Room;
 import org.junit.Before;
 import org.junit.Test;
 
 import java.util.Collection;
-import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
