+++ src/test/java/com/capgemini/molvenoresort/user/UserTest.java
@@ -1,8 +1,8 @@
 package com.capgemini.molvenoresort.user;
 
-import com.capgemini.molvenoresort.user.Employee;
-import com.capgemini.molvenoresort.user.EmployeeRole;
-import com.capgemini.molvenoresort.user.User;
+import com.capgemini.molvenoresort.enums.EmployeeRole;
+import com.capgemini.molvenoresort.models.Employee;
+import com.capgemini.molvenoresort.models.User;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
